ai Backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0
beanie==1.23.6  # ODM for MongoDB

# AI/ML dependencies
transformers==4.36.0
torch==2.1.0
tokenizers==0.15.0
accelerate==0.25.0
sentence-transformers==2.2.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# WebSocket support
websockets==12.0

# HTTP client for external APIs
httpx==0.25.2
aiohttp==3.9.1

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# CORS support
python-cors==1.7.0

# Background tasks
celery==5.3.4
redis==5.0.1
